refactor(upstream): extract pool/helpers.rs leaf helpers (lab-kvji.12.1)
Steps 1-2 of the pool.rs split: scaffold pool/ child dir and move leaf
helpers (config knobs, constants, error classification, naming, redaction,
UpstreamCachedSummary, and the prompt/resource merge/rewrite/cached_upstream_tool
helpers) into pool/helpers.rs. Re-export public surface (UpstreamCachedSummary,
in_process_upstream_name, redact_resource_uri_for_logging,
upstream_discovery_concurrency) from pool.rs; glob-import the pub(super) leaves
so residual call sites stay unqualified. Behavior unchanged. 1423 tests green.

Bead: lab-kvji.12.1
